Print current pin after every command
Previously, some commands didn't indicate which pin they were using. This was especially confusing on the first command, since there was no way to know that the default pin 2 was used. It is also a bit non-obvious that the pin used in the previous command was still in use for the new command.