Add delay before printing menu in setup()
A delay is needed for the menu to appear after opening Serial Monitor when using the Nano 33 BLE boards. I remember having a similar problem with the Uno WiFi Rev2 some time ago, but I think the problem was fixed for that board. There may be other boards with a similar requirement for a delay. The 100 ms delay is much longer than needed, but perhaps a conservative delay duration will make sure the problem is fixed for all boards, while still not being noticeable to a human.