Add note to readme re: command menu not being shown when Serial Monitor is opened
On boards without native USB (and thus won't wait for while(!Serial)) and that don't reset when Serial Monitor is opened, the menu will be printed before Serial Monitor is opened, and thus will never be seen. Of the boards I own, the ones with this problem are: Nano Every, NodeMCU-32S, and WeMos D1 Mini.